refactor: move useOverlayZIndex hook to hooks directory, fix dev-up.sh auth override
- Move useOverlayZIndex + OverlayDepthContext from components/ui/z-index.ts
  to hooks/useOverlayZIndex.ts (hooks don't belong in ui/)
- Fix NEXTAUTH_PROVIDER override: shell export doesn't reach Docker containers,
  so write to .env.dev-up file loaded by compose as optional secondary env_file
- compose.dev.yml loads langwatch/.env.dev-up (optional) after .env, so
  dev-up.sh overrides take effect while make quickstart uses .env as-is
